# MAINT: It's faster to use an astype than to multiply by 1.0