@marload 2020-07-14
> These cases are not equivalent when `x.dtype in (np.half, np.single, np.longdouble)`, or when `x` is complex

Thank You!

There is a way to deal with the exception, but I think it will reduce the readability of the code. So I think it's better to just leave it as it is. 

Is there a simple way to deal with the exceptions you are thinking of?