@seberg 2020-07-16
You would have to first use `np.result_type` and then also add `astype(dtype, copy=False)` with the copy kwarg to make it worthfile. Not sure it will be worthwhile in the end.