@mattip 2020-08-02
Closing. Thanks for the initiative, but this fix does not cover all the cases. Please reopen if you find a way to cleanly address the problems. I think if we were starting over we would not have [`scimath`](https://numpy.org/devdocs/reference/routines.emath.html?highlight=scimath#module-numpy.lib.scimath) in the first place, so maybe a better change would be to deprecate this module.